start norm
