+++ WM/includes/vm.h
@@ -30,12 +30,16 @@
 #define _FDUMP st.flag_dump
 #define _RAVNO st.flag_dump_d == st.cycles
 #define _PBF if (_RAVNO && (_FD || _FDUMP)){_PRINT_B}
-#define _CHECK_VISUALISATION {if (st.flag_visual == true){vs_main();}}
+#define _CHECK_VISUALISATION {if (st.flag_visual == true){vs_main();}_VS;}
 #define _CHECK_WAIT kareta->code >= 0x01 && kareta->code <= 0x10
 #define _LOG(ex, ex1) {if (st.log == 2){logs(ex, ex1);}}
 #define _PREPARE(ex) {st.cycles++;st.cycles_after_check++;ex = st.kareta;}
 #define _PRINT_B2 {if (st.flag_dump)print_field(32);exit(13);}
 #define _PRINT_B {if (st.flag_d){print_field(64);}_PRINT_B2;}
+#define _CHECKSSS st.check_in == MAX_CHECKS || st.live_in >= NBR_LIVE
+#define _CYCLE_D {st.cycles_to_die -= CYCLE_DELTA;st.check_in = 0;}
+#define _IS_CYCLE if  (_CHECKSSS){_CYCLE_D}
+#define _ZERO {	st.live_in = 0;	st.cycles_after_check = 0;}
 
 #define st g_gen
 
@@ -55,11 +59,18 @@ void create_field(void);
 int		count_step(uint8_t type, t_operation *oop);
 int8_t	count_size(int pos, int step);
 int		check_args(t_kareta *kareta, t_operation *oop);
-int		validate_args(t_kareta *kareta, t_operation *oop);
+bool	validate_args(t_kareta *kareta, t_operation *oop);
 void	read_argtype(t_kareta *kareta, t_operation *oop);
 t_kareta	*copy_kareta(t_kareta *kareta, int32_t addr);
 void add_elem(t_kareta **list, int pos);
 void print_list(t_kareta *list);
 void set_karetu();
+void logs(int i, int d);
+void 	parsing_argc(int argc, char **argv);
+int check_where(char *where);
+void swap(t_champ *first, t_champ *second);
 void print_info_champs();
+void	print_last_alive();
+void print_player_code();
+int32_t	bytecode_to_int32(uint8_t *bytecode, int size);
 #endif
