+++ WM/src/add_func.c
@@ -13,56 +13,6 @@ void		op_aff(t_kareta *kareta)
 		ft_printf("%c\n", (char)value);
 }
 
-void		set_field(int32_t addr, uint8_t i, int32_t size)
-{
-	int j;
-
-	j = 0;
-	while (size)
-	{
-		st.v_field[find_adress(addr + size - 1)] = i;
-		j += 8;
-		size--;
-	}
-}
-
-void		int_to_byte(int32_t addr, int32_t value, int32_t size)
-{
-	int8_t i;
-
-	i = 0;
-	while (size)
-	{
-		st.field[find_adress(addr + size - 1)] =
-			(uint8_t)((value >> i) & 0xFF);
-		i += 8;
-		size--;
-	}
-}
-
-int32_t		byte_to_int(int32_t addr, int32_t size)
-{
-	int32_t		result;
-	uint8_t		sign;
-	int			i;
-
-	i = 0;
-	result = 0;
-	sign = (st.field[find_adress(addr)] & 0x80);
-	while (size)
-	{
-		if (sign)
-			result += ((st.field[find_adress(addr + size - 1)] ^ 0xFF) <<
-					(i++ * 8));
-		else
-			result += st.field[find_adress(addr + size - 1)] << (i++ * 8);
-		size--;
-	}
-	if (sign)
-		result = ~(result);
-	return (result);
-}
-
 int			take_op(t_kareta *kareta, uint8_t i, int mod)
 {
 	t_operation	*op;
@@ -84,4 +34,39 @@ int			take_op(t_kareta *kareta, uint8_t i, int mod)
 	}
 	kareta->step += count_step(kareta->argc_types[i - 1], op);
 	return (value);
+}
+
+void swap(t_champ *first, t_champ *second)
+{
+	t_champ tmp;
+
+	tmp = *first;
+	*first = *second;
+	*second = tmp;
+}
+
+int check_where(char *where)
+{
+	int a;
+	char *tmp;
+
+	if (!where)
+		_ERROR("O_o");
+	a = ft_atoi(where);
+	tmp = ft_itoa(a);
+	if (!ft_strequ(tmp, where))
+	{
+		free(tmp);
+		_ERROR("Enter a number!");
+	}
+	free(tmp);
+	return (a);
+}
+
+int		find_adress(int i)
+{
+	i %= MEM_SIZE;
+	if (i < 0)
+		i += MEM_SIZE;
+	return (i);
 }
