+++ WM/src/read_champ.c
@@ -2,28 +2,6 @@
 #include <fcntl.h>
 #include <stdint.h>
 
-int32_t	bytecode_to_int32(uint8_t *bytecode, int size)
-{
-	int32_t	result;
-	int		sign;
-	int		i;
-
-	result = 0;
-	sign = (int)(bytecode[0] & 0x80);
-	i = 0;
-	while (size)
-	{
-		if (sign)
-			result += ((bytecode[size - 1] ^ 0xFF) << (i++ * 8));
-		else
-			result += bytecode[size - 1] << (i++ * 8);
-		size--;
-	}
-	if (sign)
-		result = ~(result);
-	return (result);
-}
-
 char	*check_name(int fd, int len)
 {
 	int		size;
@@ -31,7 +9,7 @@ char	*check_name(int fd, int len)
 
 	_ERROR_MALLOC((buffer = ft_strnew(len)));
 	size = read(fd, buffer, len);
-	_READING(size < len, "ERROR FILE")
+	_READING(size < len, "Error Name/Comment")
 	return (buffer);
 }
 
@@ -41,7 +19,7 @@ int32_t	check_int(int fd)
 	uint8_t	buffer[4];
 
 	size = read(fd, &buffer, 4);
-	_READING(size == -1 || size < 4, "ERROR FILE");
+	_READING(size == -1 || size < 4, "Size Error, U're so FAT");
 	return (bytecode_to_int32(buffer, 4));
 }
 
@@ -55,7 +33,7 @@ uint8_t	*check_code(int fd, int len)
 	size = read(fd, buffer, len);
 	_READING(size == -1, "SIZE ERROR")
 	_READING(size < (ssize_t)len || read(fd, &byte, 1) != 0,
-			"Error file");
+			"Invalid Exucation Code");
 	return (buffer);
 }
 
@@ -65,12 +43,12 @@ void read_champ(int a, char *file_name)
 
 	ft_strcat(g_gen.champ[a].file_name, file_name);
 	fd = open(file_name, O_RDONLY);
-	_READING(check_int(fd) != COREWAR_EXEC_MAGIC, "ZA SHO?");
+	_READING(check_int(fd) != COREWAR_EXEC_MAGIC, "Magic isn't Power :(");
 	g_gen.champ[a].name = check_name(fd, PROG_NAME_LENGTH);
-	_READING(check_int(fd) != 0, "ERROR FILE");
+	_READING(check_int(fd) != 0, "Missing Zero Byte");
 	_READING (((g_gen.champ[a].length = check_int(fd)) < 0
-		|| g_gen.champ[a].length > CHAMP_MAX_SIZE), "ERROR_SIZE");
+		|| g_gen.champ[a].length > CHAMP_MAX_SIZE), "Error Champ Size");
 	g_gen.champ[a].comment = check_name(fd, COMMENT_LENGTH);
-	_READING(check_int(fd) != 0, "ERROR");
+	_READING(check_int(fd) != 0, "Missing Zero Byte");
 	g_gen.champ[a].algo = check_code(fd, g_gen.champ[a].length);
 }
